Move to Tensor.device
